Fix blank WebView: GenPanel crash + reliable single-file UI loading
The GUI WebView rendered blank since the SA3 GenPanel rewrite. Two layered
causes, found by capturing the in-page JS error:

1. ROOT CAUSE — GenPanel's GenBody crashed on first render:
   `(rl.nl ?? 0.45).toFixed(2)` threw "toFixed is not a function" because
   `nl` arrives as a STRING (ValueTree numbers become strings across the
   .tracktionedit XML save/reload). With no error boundary, the thrown render
   unmounted the ENTIRE React tree → blank window. The earlier "looked great"
   screenshot was the pre-SA3 fake panel, so this never showed before.
   - GenPanel.tsx: coerce nl/seed with Number() (defensive at the seam).
   - MoshOps.cpp: serialize renderLayer nl/seed/colour-value as real numbers
     in the snapshot, so the contract is correctly typed at the source.
   - main.tsx: add an ErrorBoundary — a single component throwing now shows the
     error in-app instead of blanking everything.

2. LOADING — JUCE's resource-provider custom scheme does not execute external
   `<script type=module crossorigin>` (nor apply external crossorigin CSS), so
   the bundle could silently never run. Switch the Vite build to
   vite-plugin-singlefile (inlines JS+CSS into one index.html); inline scripts/
   styles need no fetch and always run. Removes the runtime C++ HTML-rewriting
   hack from WebBridge.cpp.

Also:
- WebViewShell::load() defers goToURL until after Main.cpp wires the command/
  snapshot/event handlers, so the page can't call native fns before they bind.
- BuildUI.cmake now tracks ui/src so a UI-source change triggers a rebuild
  (previously only package.json/vite.config.ts were watched). Re-enable with
  -DMOSH_BUILD_UI=ON (default ON).

Verified: full UI renders (bare + --demo5), Mosh --selftest 89/89.

+++ src/moshops/MoshOps.cpp
@@ -1183,20 +1183,20 @@ juce::var MoshOps::clipToVar (te::Clip& c)
         r->setProperty ("status", rl[ids::status]);
         r->setProperty ("adapter", rl[ids::modelAdapter]);
         r->setProperty ("mode", rl[ids::mode]);
-        r->setProperty ("seed", rl[ids::seed]);
+        r->setProperty ("seed", (int) rl[ids::seed]);
         r->setProperty ("userKept", rl[ids::userKept]);
         r->setProperty ("hasArtifact", juce::File (rl[ids::cacheArtifact].toString()).existsAsFile());
         if (auto params = rl.getChildWithName (ids::PARAMS); params.isValid())
         {
             r->setProperty ("prompt", params[ids::prompt]);
-            r->setProperty ("nl", params[ids::nl]);
+            r->setProperty ("nl", (double) params[ids::nl]);
             Array<var> colors;
             if (auto cs = params.getChildWithName (ids::COLORS); cs.isValid())
                 for (int i = 0; i < cs.getNumChildren(); ++i)
                 {
                     auto* co = new DynamicObject();
                     co->setProperty ("name", cs.getChild (i)[ids::name]);
-                    co->setProperty ("value", cs.getChild (i)[ids::value]);
+                    co->setProperty ("value", (double) cs.getChild (i)[ids::value]);
                     colors.add (var (co));
                 }
             r->setProperty ("colors", colors);
